detect schema-changes for db-backups
in case the schema for a test-db is changed, an outdated database would
be restored from the backup since the hash in the backup-filename only
contains the list of fixtures to be loaded.
This patch adds the complete metadata to the hash to make sure the hash
will change when anything in the schema changes.