Remove cached databases if fixtures are updated
Determine if the fixtures that define a database backup have been
modified since the database backup was made. This is done by comparing
the time last modified (`filemtime`) of all the individual fixtures with
the backup SQLite database.

If the SQLite database time of last modification is older than the
fixture definitions the database is assumed to be out of date and then
is regenerated.