Handle Symfony disabling the validator component
Symfony now allows the user to disable the validator component, which
results in the validator service being undefined. This then causes
errors on container compilation as we expected this to always exist.

To fix this add a compiler pass that removes our wrapper around
validator if the validator itself is undefined.